feat: 3차 세미나 실습 구현 및 도메인형 패키지 구조로 변경
