--- src/main/java/org/sopt/common/CommonErrorCode.java
@@ -1,28 +1,0 @@
-package org.sopt.common;
-
-import org.springframework.http.HttpStatus;
-
-public enum CommonErrorCode implements ErrorCode {
-
-    BAD_REQUEST(HttpStatus.BAD_REQUEST, "COMMON_400", "잘못된 요청입니다."),
-    INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "COMMON_500", "서버 내부 오류가 발생했습니다.");
-
-    private final HttpStatus httpStatus;
-    private final String code;
-    private final String message;
-
-    CommonErrorCode(HttpStatus httpStatus, String code, String message) {
-        this.httpStatus = httpStatus;
-        this.code = code;
-        this.message = message;
-    }
-
-    @Override
-    public HttpStatus getHttpStatus() { return httpStatus; }
-
-    @Override
-    public String getCode() { return code; }
-
-    @Override
-    public String getMessage() { return message; }
-}
