--- src/main/java/org/sopt/common/CustomException.java
@@ -1,13 +1,0 @@
-package org.sopt.common;
-
-public class CustomException extends RuntimeException {
-
-    private final ErrorCode errorCode;
-
-    public CustomException(ErrorCode errorCode) {
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-    }
-
-    public ErrorCode getErrorCode() { return errorCode; }
-}
