--- src/main/java/org/sopt/common/GlobalExceptionHandler.java
@@ -1,25 +1,0 @@
-package org.sopt.common;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    // CustomException을 상속한 모든 예외 처리
-    @ExceptionHandler(CustomException.class)
-    public ResponseEntity<ApiResponse<Void>> handleCustomException(CustomException e) {
-        return ResponseEntity
-                .status(e.getErrorCode().getHttpStatus())
-                .body(ApiResponse.error(e.getErrorCode()));
-    }
-
-    // 그 외 예외 처리
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<ApiResponse<Void>> handleException(Exception e) {
-        return ResponseEntity
-                .status(CommonErrorCode.INTERNAL_SERVER_ERROR.getHttpStatus())
-                .body(ApiResponse.error(CommonErrorCode.INTERNAL_SERVER_ERROR));
-    }
-}
