--- src/main/java/org/sopt/common/SuccessCode.java
@@ -1,9 +1,0 @@
-package org.sopt.common;
-
-import org.springframework.http.HttpStatus;
-
-public interface SuccessCode {
-    HttpStatus getHttpStatus();
-    String getCode();
-    String getMessage();
-}
