+++ src/main/java/org/sopt/domain/post/dto/response/DeletePostResponse.java
@@ -1,0 +1,7 @@
+package org.sopt.domain.post.dto.response;
+
+// 게시글 삭제 응답
+public record DeletePostResponse(
+        Long id
+) {
+}
