+++ src/main/java/org/sopt/domain/post/exception/PostException.java
@@ -1,0 +1,10 @@
+package org.sopt.domain.post.exception;
+
+import org.sopt.global.CustomException;
+
+public class PostException extends CustomException {
+
+    public PostException(PostErrorCode errorCode) {
+        super(errorCode);
+    }
+}
