+++ src/main/java/org/sopt/domain/post/repository/PostRepository.java
@@ -1,0 +1,9 @@
+package org.sopt.domain.post.repository;
+
+import org.sopt.domain.post.entity.Post;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PostRepository extends JpaRepository<Post, Long> {
+}
