+++ src/main/java/org/sopt/domain/user/entity/User.java
@@ -1,0 +1,27 @@
+package org.sopt.domain.user.entity;
+
+import jakarta.persistence.*;
+
+@Entity
+@Table(name = "users")
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String nickname;
+
+    private String email;
+
+    protected User() { }
+
+    public User(String nickname, String email) {
+        this.nickname = nickname;
+        this.email = email;
+    }
+
+    public Long getId() { return this.id; }
+    public String getNickname() { return this.nickname; }
+    public String getEmail() { return this.email; }
+}
