+++ src/main/java/org/sopt/domain/user/exception/UserException.java
@@ -1,0 +1,10 @@
+package org.sopt.domain.user.exception;
+
+import org.sopt.global.CustomException;
+
+public class UserException extends CustomException {
+
+    public UserException(UserErrorCode errorCode) {
+        super(errorCode);
+    }
+}
