+++ src/main/java/org/sopt/domain/user/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package org.sopt.domain.user.repository;
+
+import org.sopt.domain.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String email);
+}
