--- src/main/java/org/sopt/dto/request/CreatePostRequest.java
@@ -1,9 +1,0 @@
-package org.sopt.dto.request;
-
-// 게시글 작성 요청
-public record CreatePostRequest(
-        String title,
-        String content,
-        String author
-) {
-}
