--- src/main/java/org/sopt/dto/response/DeletePostResponse.java
@@ -1,7 +1,0 @@
-package org.sopt.dto.response;
-
-// 게시글 수정 응답
-public record DeletePostResponse(
-        Long id
-) {
-}
