--- src/main/java/org/sopt/dto/response/PostResponse.java
@@ -1,22 +1,0 @@
-package org.sopt.dto.response;
-
-import org.sopt.domain.Post;
-
-// 게시글 단건 조회 응답
-public record PostResponse(
-        Long id,
-        String title,
-        String content,
-        String author,
-        String createdAt
-) {
-    public static PostResponse from(Post post) {
-        return new PostResponse(
-                post.getId(),
-                post.getTitle(),
-                post.getContent(),
-                post.getAuthor(),
-                post.getCreatedAt()
-        );
-    }
-}
