--- src/main/java/org/sopt/exception/PostException.java
@@ -1,10 +1,0 @@
-package org.sopt.exception;
-
-import org.sopt.common.CustomException;
-
-public class PostException extends CustomException {
-
-    public PostException(PostErrorCode errorCode) {
-        super(errorCode);
-    }
-}
