+++ src/main/java/org/sopt/global/CustomException.java
@@ -1,0 +1,15 @@
+package org.sopt.global;
+
+import org.sopt.global.code.ErrorCode;
+
+public class CustomException extends RuntimeException {
+
+    private final ErrorCode errorCode;
+
+    public CustomException(ErrorCode errorCode) {
+        super(errorCode.getMessage());
+        this.errorCode = errorCode;
+    }
+
+    public ErrorCode getErrorCode() { return errorCode; }
+}
