+++ src/main/java/org/sopt/global/code/ErrorCode.java
@@ -1,0 +1,9 @@
+package org.sopt.global.code;
+
+import org.springframework.http.HttpStatus;
+
+public interface ErrorCode {
+    HttpStatus getHttpStatus();
+    String getCode();
+    String getMessage();
+}
