+++ src/main/java/org/sopt/global/code/SuccessCode.java
@@ -1,0 +1,9 @@
+package org.sopt.global.code;
+
+import org.springframework.http.HttpStatus;
+
+public interface SuccessCode {
+    HttpStatus getHttpStatus();
+    String getCode();
+    String getMessage();
+}
